Match registries allow list by regex
Related to STONEBLD-1831

Since this commit, the every line within the registries allow list
are changed to regular expression. re.match is used for pattern
matching, which causes the parent image reference is matched from the
beginning. This is flexible for configuring the allow list to match more
specifically.

This change is compatible with existing allow list. If necessary, the
migration would be slight changes by converting "." to "\.".

Signed-off-by: Chenxiong Qi <cqi@redhat.com>